o/snapstate: check disk space w/o store if possible
The disk space checks were obtaining the size of the snaps' prerequisites and
base via a call to the store. However, an InstallPathMany should allow the
user to install without calls to the store, if all requirements are
available locally. This change makes it so that, if the required bases and
prereqs are included in the install, those are used for the disk check instead
of gotten through the store. Note that for Install/UpdateMany with
prerequisites the already downloaded prereqs are also used, instead of
redownloaded.