aspects: match requests that prefix patterns
Previously, requests had to match an aspect's access pattern precisely
(not including placeholders). This relaxes the request matching to first
look for an exact match and then allow a match if the request is a
prefix of the pattern. The result is returned in a nested namespace that
it was defined on (in the access rule) but that is also flattened by
the request prefix. For example, if "snaps.snapd.status" is a rule and
"snaps.snapd" is requested, we could get `{ "snaps.snapd": { "status": "active" }}`.
So the request string cuts through the result namespace.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>