o/devicestate: pin time in install mode tests
Some tests in this suite were flaky because the time difference between
the seed and model creation and signing. If the second signing occurred
in a different second then the signatures would differ and the test
would fail. An alternative approach would be to not regenerate the model
at all but it would require a larger refactor of the test helpers.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>